pyproject.toml
[project]
name = "deep-research-anything"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
packages = [{include = "*", from = "deep_research_anything"}]
readme = "README.md"
requires-python = ">=3.11,<4.0"

dependencies = [
    "litellm (>=1.63.3,<2.0.0)",
    "firecrawl-py (>=1.12.0,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "streamlit (>=1.43.1,<2.0.0)",
    "llama-index (>=0.12.23,<0.13.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"